separation of alibaba requests
